requirements.txt
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.5.3
asgiref==3.8.1
assemblyai==0.36.0
babel==2.17.0
bcrypt==4.1.2
Beaker==1.12.1
beautifulsoup4==4.12.3
blivet==3.11.0
#blivet-gui==2.6.0
boto3==1.36.12
botocore==1.36.12
#Brlapi==0.8.5
Brotli==1.1.0
cachetools==5.5.1
capstone==5.0.1
certifi==2025.1.31
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
colored-traceback==0.3.0
crispy-bootstrap4==2024.10
crypt_r==3.13.1
cryptography==43.0.0
cssselect2==0.7.0
#cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
distlib==0.3.9
distro==1.9.0
Django==5.1.5
django-allauth==65.3.1
django-cors-headers==4.6.0
django-countries==7.6.1
django-crispy-forms==2.3
django-weasyprint==2.3.1
djangorestframework==3.15.2
#dnf==4.22.0
dnspython==2.6.1
et_xmlfile==2.0.0
#fedora-third-party==0.10
file-magic==0.4.0
filelock==3.17.0
fonttools==4.55.8
#fros==1.1
google-auth==2.38.0
google-genai==1.0.0
#gpg==1.23.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.9.0
idna==3.7
intervaltree==3.1.0
iso639==0.1.4
jmespath==1.0.1
langtable==0.0.68
libcomps==0.1.21.post1
#libdnf==0.73.4
louis==1.3
lxml==5.2.1
Mako==1.2.3
Markdown==3.7
MarkupSafe==2.1.5
#nftables==0.1
numpy==2.2.1
oauthlib==3.2.2
olefile==0.47
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
paramiko==3.5.0
Paste==3.10.1
pexpect==4.9.0
pid==2.2.3
pillow==11.0.0
platformdirs==4.3.6
plumbum==1.7.2
ply==3.11
productmd==1.43
progressbar2==3.53.2
psutil==5.9.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pwntools==4.13.1
#pwquality==1.4.5
pyasn1==0.6.0
pyasn1_modules==0.4.1
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.11.0
pyelftools==0.31
pyenchant==3.2.2
pygit2==1.17.0
Pygments==2.18.0
#PyGObject==3.48.2
PyJWT==2.10.1
pykickstart==3.58
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparted==3.13.0
pyphen==0.17.2
pyserial==3.5
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-decouple==3.8
#python-meh==0.52
python-pam==2.0.2
python-utils==3.7.0
pytz==2024.2
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
regex==2024.9.11
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
ROPGadget==7.4
#rpm==4.20.0
rpmautospec==0.7.3
rpmautospec-core==0.1.5
rpyc==6.0.1
rsa==4.9
s3transfer==0.11.2
#selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
#sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
#setools==4.5.1
setproctitle==1.3.3
setuptools==75.8.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
#sos==4.8.2
soupsieve==2.6
sqlparse==0.5.3
systemd-python==235
Tempita==0.5.2
tiktoken==0.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
#ufw==0.35
unicorn==2.0.1.post1
unix_ar==0.2.1
urllib3==1.26.20
virtualenv==20.29.1
weasyprint==64.0
webencodings==0.5.1
websockets==14.2
xkbregistry==0.3
xlrd==2.0.1
youtube_dl==2021.12.17
zopfli==0.2.3.post1
zstandard==0.23.0
